refactor(core): rename generateFiles to templates
feat(core/templates): add getTemplate
